+++ src/editor/DebugMode.h
@@ -36,9 +36,11 @@ public:
 
     void renderEntitiesPanel(SceneManager &sceneManager);
 
+    void renderScenesPanel(SceneManager &sceneManager);
+
 private:
-    bool _showDemoWindow = true;
-    bool _showSimpleWindow = true;
+    bool _showDemoWindow = false;
+    bool _showSimpleWindow = false;
     bool _showAnotherWindow = false;
     ImVec4 _clearColor = ImVec4(0.45f, 0.55f, 0.60f, 1.00f);
 };